+++ ch_tools/chadmin/cli/object_storage_group.py
@@ -1,28 +1,13 @@
 from datetime import timedelta
-from tempfile import TemporaryFile
 from typing import Optional
 
-import click
 from click import Context, group, option, pass_context
 from humanfriendly import format_size
 
-from ch_tools.chadmin.internal.object_storage import (
-    ObjListItem,
-    cleanup_s3_object_storage,
-)
-from ch_tools.chadmin.internal.object_storage.utils import (
-    DEFAULT_GUARD_INTERVAL,
-    get_orphaned_objects_query,
-    get_remote_data_paths_table,
-    get_traverse_shadow_settings,
-    traverse_object_storage,
-)
-from ch_tools.chadmin.internal.utils import execute_query
-from ch_tools.common import logging
 from ch_tools.common.cli.formatting import print_response
 from ch_tools.common.cli.parameters import TimeSpanParamType
 from ch_tools.common.clickhouse.config import get_clickhouse_config
-from ch_tools.common.clickhouse.config.storage_configuration import S3DiskConfiguration
+from ch_tools.common.commands.clean_object_storage import DEFAULT_GUARD_INTERVAL, clean
 
 # Use big enough timeout for stream HTTP query
 STREAM_TIMEOUT = 10 * 60
@@ -98,7 +83,7 @@ def object_storage_group(ctx: Context, disk_name: str) -> None:
     "--keep-paths",
     "keep_paths",
     is_flag=True,
-    help=("Do not delete collected paths of objects from object storage."),
+    help=("Do not delete the collected paths from the local table, when the command completes."),
 )
 @option(
     "--use-saved-list",
@@ -121,100 +106,18 @@ def clean_command(
     """
     Clean orphaned S3 objects.
     """
-    if from_time is not None and to_time <= from_time:
-        raise click.BadParameter(
-            "'to_time' parameter must be greater than 'from_time'",
-            param_hint="--from-time",
-        )
-
-    disk_conf: S3DiskConfiguration = ctx.obj["disk_configuration"]
-    config = ctx.obj["config"]["object_storage"]["clean"]
-
-    listing_table = f"{config['listing_table_database']}.{config['listing_table_prefix']}{disk_conf.name}"
-    # Create listing table for storing paths from object storage
-    try:
-        execute_query(
-            ctx,
-            f"CREATE TABLE IF NOT EXISTS {listing_table} (obj_path String, obj_size UInt64) ENGINE MergeTree ORDER BY obj_path SETTINGS storage_policy = '{config['storage_policy']}'",
-        )
-        _clean_object_storage(
-            ctx,
-            object_name_prefix,
-            from_time,
-            to_time,
-            on_cluster,
-            cluster_name,
-            dry_run,
-            listing_table,
-            use_saved_list,
-        )
-    finally:
-        if not keep_paths:
-            execute_query(
-                ctx, f"DROP TABLE IF EXISTS {listing_table} SYNC", format_=None
-            )
-
-
-def _clean_object_storage(
-    ctx: Context,
-    object_name_prefix: str,
-    from_time: Optional[timedelta],
-    to_time: timedelta,
-    on_cluster: bool,
-    cluster_name: str,
-    dry_run: bool,
-    listing_table: str,
-    use_saved_list: bool,
-) -> None:
-    """
-    Delete orphaned objects from object storage.
-    """
-    disk_conf: S3DiskConfiguration = ctx.obj["disk_configuration"]
-    prefix = object_name_prefix or disk_conf.prefix
-
-    if not use_saved_list:
-        logging.info(
-            f"Collecting objects... (Disk: '{disk_conf.name}', Endpoint '{disk_conf.endpoint_url}', Bucket: '{disk_conf.bucket_name}', Prefix: '{disk_conf.prefix}')",
-        )
-        counter = traverse_object_storage(
-            ctx, listing_table, from_time, to_time, prefix
-        )
-        logging.info("Collected {} objects", counter)
-
-    remote_data_paths_table = get_remote_data_paths_table(on_cluster, cluster_name)
-    settings = get_traverse_shadow_settings(ctx)
-    antijoin_query = get_orphaned_objects_query(
-        listing_table, remote_data_paths_table, disk_conf, settings
-    )
-    logging.info("Antijoin query: {}", antijoin_query)
-
-    if dry_run:
-        logging.info("Counting orphaned objects...")
-    else:
-        logging.info("Deleting orphaned objects...")
-
-    deleted = 0
-    total_size = 0
-    with TemporaryFile() as keys_file:
-        with execute_query(
-            ctx, antijoin_query, stream=True, format_="TabSeparated"
-        ) as resp:
-            # Save response to the file by chunks
-            for chunk in resp.iter_content(chunk_size=8192):
-                keys_file.write(chunk)
-
-        keys_file.seek(0)  # rewind file pointer to the beginning
-
-        keys = (
-            ObjListItem.from_tab_separated(line.decode().strip()) for line in keys_file
-        )
-        deleted, total_size = cleanup_s3_object_storage(disk_conf, keys, dry_run)
-
-    logging.info(
-        f"{'Would delete' if dry_run else 'Deleted'} {deleted} objects with total size {format_size(total_size, binary=True)} from bucket [{disk_conf.bucket_name}] with prefix {prefix}",
+    deleted, total_size = clean(
+        ctx,
+        object_name_prefix,
+        from_time,
+        to_time,
+        on_cluster,
+        cluster_name,
+        dry_run,
+        keep_paths,
+        use_saved_list
     )
     _print_response(ctx, dry_run, deleted, total_size)
-
 
 def _print_response(ctx: Context, dry_run: bool, deleted: int, total_size: int) -> None:
     """
