+++ ch_tools/monrun_checks/ch_orphaned_objects.py
@@ -1,30 +1,17 @@
 from datetime import timedelta
-from tempfile import TemporaryFile
 from typing import Optional
 
 import click
 
-from ch_tools.chadmin.internal.object_storage.obj_list_item import ObjListItem
-from ch_tools.chadmin.internal.object_storage.s3_cleanup import (
-    cleanup_s3_object_storage,
-)
-from ch_tools.chadmin.internal.object_storage.utils import (
-    DEFAULT_GUARD_INTERVAL,
-    get_orphaned_objects_query,
-    get_remote_data_paths_table,
-    get_traverse_shadow_settings,
-    traverse_object_storage,
-)
-from ch_tools.chadmin.internal.utils import execute_query
 from ch_tools.common.cli.parameters import TimeSpanParamType
 from ch_tools.common.clickhouse.config import get_clickhouse_config
-from ch_tools.common.clickhouse.config.storage_configuration import S3DiskConfiguration
+from ch_tools.common.commands.clean_object_storage import DEFAULT_GUARD_INTERVAL, clean
 from ch_tools.common.result import CRIT, OK, WARNING, Result
 
 DRY_RUN = True
 ON_CLUSTER = True
 CLUSTER_NAME = "{cluster}"
-DIST_NAME = "object_storage"
+DISK_NAME = "object_storage"
 OBJECT_NAME_PREFIX = None
 
 
@@ -33,7 +20,7 @@ OBJECT_NAME_PREFIX = None
     "--keep-paths",
     "keep_paths",
     is_flag=True,
-    help=("Do not delete collected paths of objects from object storage."),
+    help=("Do not delete the collected paths from the local table, when the command completes."),
 )
 @click.option(
     "--use-saved-list",
@@ -79,36 +66,19 @@ def orphaned_objects_command(
     ch_config = get_clickhouse_config(ctx)
     ctx.obj[
         "disk_configuration"
-    ] = ch_config.storage_configuration.s3_disk_configuaration(DIST_NAME)
+    ] = ch_config.storage_configuration.s3_disk_configuaration(DISK_NAME)
 
-    if from_time is not None and from_time <= to_time:
-        raise click.BadParameter(
-            "'to_time' parameter must be greater than 'from_time'",
-            param_hint="--from-time",
-        )
-
-    disk_conf: S3DiskConfiguration = ctx.obj["disk_configuration"]
-    config = ctx.obj["config"]["object_storage"]["clean"]
-
-    listing_table = f"{config['listing_table_database']}.{config['listing_table_prefix']}{disk_conf.name}"
-    # Create listing table for storing paths from object storage
-    try:
-        execute_query(
-            ctx,
-            f"CREATE TABLE IF NOT EXISTS {listing_table} (obj_path String, obj_size UInt64) ENGINE MergeTree ORDER BY obj_path SETTINGS storage_policy = '{config['storage_policy']}'",
-        )
-        total_size = _get_total_size(
-            ctx,
-            from_time,
-            to_time,
-            listing_table,
-            use_saved_list,
-        )
-    finally:
-        if not keep_paths:
-            execute_query(
-                ctx, f"DROP TABLE IF EXISTS {listing_table} SYNC", format_=None
-            )
+    _, total_size = clean(
+        ctx,
+        OBJECT_NAME_PREFIX,
+        from_time,
+        to_time,
+        ON_CLUSTER,
+        CLUSTER_NAME,
+        DRY_RUN,
+        keep_paths,
+        use_saved_list
+    )
 
     msg = f"Total size: {total_size}"
     if total_size >= crit:
@@ -116,44 +86,3 @@ def orphaned_objects_command(
     if total_size >= warn:
         return Result(WARNING, msg)
     return Result(OK, msg)
-
-
-def _get_total_size(
-    ctx: click.Context,
-    from_time: Optional[timedelta],
-    to_time: timedelta,
-    listing_table: str,
-    use_saved_list: bool,
-) -> int:
-    """
-    Delete orphaned objects from object storage.
-    """
-    disk_conf: S3DiskConfiguration = ctx.obj["disk_configuration"]
-    prefix = OBJECT_NAME_PREFIX or disk_conf.prefix
-
-    if not use_saved_list:
-        _ = traverse_object_storage(ctx, listing_table, from_time, to_time, prefix)
-
-    remote_data_paths_table = get_remote_data_paths_table(ON_CLUSTER, CLUSTER_NAME)
-    settings = get_traverse_shadow_settings(ctx)
-    antijoin_query = get_orphaned_objects_query(
-        listing_table, remote_data_paths_table, disk_conf, settings
-    )
-
-    total_size = 0
-    with TemporaryFile() as keys_file:
-        with execute_query(
-            ctx, antijoin_query, stream=True, format_="TabSeparated"
-        ) as resp:
-            # Save response to the file by chunks
-            for chunk in resp.iter_content(chunk_size=8192):
-                keys_file.write(chunk)
-
-        keys_file.seek(0)  # rewind file pointer to the beginning
-
-        keys = (
-            ObjListItem.from_tab_separated(line.decode().strip()) for line in keys_file
-        )
-        _, total_size = cleanup_s3_object_storage(disk_conf, keys, DRY_RUN)
-
-    return total_size
